refactor: Fix paths after rebase
